Update jacoco.gradle
The reporting part doesn't run if there's no data file.

Before, if you ran this plugin in a project without any tests or with tests skipped, it would fail because it tried to generate a report from a missing data file. This adds a simple check for file existence and exits the reporting early with an INFO-level log to let the user know what's going on.